webserver/requirements.txt
# 웹서버 및 크론탭 스케줄러 필수 패키지

# 웹 프레임워크
flask>=2.3.0
flask-cors>=4.0.0

# 스케줄링
schedule>=1.2.0

# 텔레그램 봇
python-telegram-bot>=20.0

# 유틸리티
python-dotenv>=1.0.0
requests>=2.28.0
pandas>=1.5.0
numpy>=1.21.0

# 설치 방법:
# pip install -r requirements.txt
